requirements.txt
Flask==3.1.0
google-generativeai==0.8.3
python-dotenv==1.0.1
MarkupSafe==3.0.2
Werkzeug==3.1.4
